+++ src/fulcro_spec/core.cljs
@@ -1,0 +1,10 @@
+(ns fulcro-spec.core
+  (:require-macros
+    [fulcro-spec.core])
+  (:require
+    [cljs.test :include-macros true]
+    [fulcro-spec.assertions]
+    [fulcro-spec.async]
+    [fulcro-spec.runner] ;;side effects
+    [fulcro-spec.selectors]
+    [fulcro-spec.stub]))